+++ web/documentserver-example/php-laravel/app/Http/Controllers/EditorController.php
@@ -117,6 +117,7 @@ class EditorController extends Controller
                     userDirectory: $request->ip(),
                     user: $user['id'],
                     withSample: $withSample,
+                    lang: $lang,
                 ));
 
             return redirect()