+++ web/documentserver-example/php-laravel/app/UseCases/Document/Create/CreateDocumentFromTemplateRequest.php
@@ -25,5 +25,6 @@ class CreateDocumentFromTemplateRequest
         public string $userDirectory,
         public string $user,
         public bool $withSample,
+        public string $lang,
     ) {}
 }